feat(bench/§143-144): add locomo common benchmark and answer guards
Adds the shared LoCoMo runner/Agentmemory adapter and improves timestamp anchoring plus extraction guards so benchmark failures map to actionable retrieval quality signals.
